+++ CHANGELOG.md
@@ -8,8 +8,8 @@
 - Changed `Map` `setPosition` signature to `(latlng, zoom, angle)`.
 - Changed `Map` `panTo` and `zoomPanTo` to accept `latlng` instead of `lat, lng`.
 - Made all `Map` `zoomPanTo` arguments except `latlng` optional.
-- Changed `Transform` API: `z` -> `zoom`, `zoom` -> `tileZoom`; `zoomAroundTo` now accepts `zoom` instead of `scale`;
-  `x` and `y` return the center point instead of top/left;
+- Changed `Transform` API: `z` renamed to `zoom`, `zoom` to `tileZoom`; `zoomAroundTo` now accepts `zoom` instead of `scale`;
+  `x` and `y` return the center point instead of top/left; `lat`/`lng` is now `center` (LatLng)
 - Added `tileSize` option for tile sources.
 - zoom levels fixed to match standard 256px tile maps. Each zoom level is 1 bigger than before.